requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.7
django-appconf==1.0.5
django-cors-headers==4.3.0
django-staticfiles==1.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
inflection==0.5.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
Pillow==10.1.0
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.1
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.12.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
